remove celer solver and kept celer fixed and free solver
